restructure code a bit
