+++ src/data-generator/auth.ts
@@ -1,16 +1,7 @@
-import faker from '@faker-js/faker'
 import fetch from 'node-fetch'
 import { AUTH_API_HOST } from './constants'
 import { VERIFICATION_CODE } from './index'
-
-export type User = {
-  username: string
-  password: string
-  token: string
-  stillInUse: boolean
-  primaryOfficeId: string
-  isSystemUser: boolean
-}
+import { User } from './users'
 
 export async function getToken(
   username: string,
@@ -76,143 +67,6 @@ export async function getTokenForSystemClient(
   const res = await authenticateResponse.json()
 
   return res.token
-}
-
-export async function createUser(
-  token: string,
-  primaryOfficeId: string,
-  overrides: Record<string, string>
-) {
-  const firstName = faker.name.firstName()
-  const familyName = faker.name.lastName()
-
-  const user = {
-    name: [
-      {
-        use: 'en',
-        firstNames: firstName,
-        familyName: familyName
-      }
-    ],
-    identifier: [
-      {
-        system: 'NATIONAL_ID',
-        value: faker.datatype
-          .number({ min: 100000000, max: 999999999 })
-          .toString()
-      }
-    ],
-    username:
-      firstName.toLocaleLowerCase() + '.' + familyName.toLocaleLowerCase(),
-    mobile: faker.phone.phoneNumber(),
-    email: faker.internet.email(),
-    primaryOffice: primaryOfficeId,
-    ...overrides
-  }
-
-  const createUserRes = await fetch('http://localhost:7070/graphql', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${token}`,
-      'x-correlation': `createuser-${firstName}-${familyName}`
-    },
-    body: JSON.stringify({
-      query: `
-      mutation createOrUpdateUser($user: UserInput!) {
-        createOrUpdateUser(user: $user) {
-          username
-          id
-        }
-      }
-    `,
-      variables: { user }
-    })
-  })
-
-  const { data } = (await createUserRes.json()) as {
-    data: { createOrUpdateUser: { username: string; id: string } }
-  }
-  const userToken = await getToken(data.createOrUpdateUser.username, 'test')
-
-  const res = await fetch('http://localhost:7070/graphql', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${userToken}`,
-      'x-correlation': `createuser-${firstName}-${familyName}`
-    },
-    body: JSON.stringify({
-      query: `
-      mutation activateUser($userId: String!, $password: String!, $securityQNAs: [SecurityQuestionAnswer]!) {
-        activateUser(userId: $userId, password: $password, securityQNAs: $securityQNAs)
-      }
-    `,
-      variables: {
-        userId: data.createOrUpdateUser.id,
-        password: 'test',
-        securityQNAs: []
-      }
-    })
-  })
-
-  if (res.status !== 200) {
-    console.log(res)
-
-    throw new Error('Could not activate user')
-  }
-
-  return {
-    ...data.createOrUpdateUser,
-    token: userToken,
-    primaryOfficeId,
-    stillInUse: true,
-    password: 'test',
-    isSystemUser: false
-  }
-}
-
-export async function createSystemClient(
-  token: string,
-  officeId: string,
-  scope: 'HEALTH' | 'NATIONAL_ID' | 'EXTERNAL_VALIDATION' | 'AGE_CHECK'
-): Promise<User> {
-  const systemAdmin = await createUser(token, officeId, {
-    role: 'LOCAL_SYSTEM_ADMIN'
-  })
-
-  const createUserRes = await fetch(
-    'http://localhost:3030/registerSystemClient',
-    {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${systemAdmin.token}`,
-        'x-correlation': `create-system-scope`
-      },
-      body: JSON.stringify({ scope })
-    }
-  )
-
-  const credentials: {
-    client_id: string
-    client_secret: string
-    sha_secret: string
-  } = await createUserRes.json()
-
-  const systemToken = await getTokenForSystemClient(
-    credentials.client_id,
-    credentials.client_secret
-  )
-
-  return {
-    token: systemToken,
-    username: credentials.client_id,
-    password: credentials.client_secret,
-    stillInUse: true,
-    primaryOfficeId: officeId,
-    isSystemUser: true
-  }
 }
 
 export function readToken(token: string) {
