+++ src/data-generator/users.ts
@@ -1,0 +1,214 @@
+import faker from '@faker-js/faker'
+import { log } from './util'
+import { Location } from './location'
+import { getFacilities } from './index'
+import fetch from 'node-fetch'
+import { getToken, getTokenForSystemClient } from './auth'
+
+export type User = {
+  username: string
+  password: string
+  token: string
+  stillInUse: boolean
+  primaryOfficeId: string
+  isSystemUser: boolean
+}
+
+type Config = {
+  fieldAgents: number
+  hospitalFieldAgents: number
+  registrationAgents: number
+  localRegistrars: number
+}
+
+export async function createUser(
+  token: string,
+  primaryOfficeId: string,
+  overrides: Record<string, string>
+) {
+  const firstName = faker.name.firstName()
+  const familyName = faker.name.lastName()
+
+  const user = {
+    name: [
+      {
+        use: 'en',
+        firstNames: firstName,
+        familyName: familyName
+      }
+    ],
+    identifier: [
+      {
+        system: 'NATIONAL_ID',
+        value: faker.datatype
+          .number({ min: 100000000, max: 999999999 })
+          .toString()
+      }
+    ],
+    username:
+      firstName.toLocaleLowerCase() + '.' + familyName.toLocaleLowerCase(),
+    mobile: faker.phone.phoneNumber(),
+    email: faker.internet.email(),
+    primaryOffice: primaryOfficeId,
+    ...overrides
+  }
+
+  const createUserRes = await fetch('http://localhost:7070/graphql', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${token}`,
+      'x-correlation': `createuser-${firstName}-${familyName}`
+    },
+    body: JSON.stringify({
+      query: `
+      mutation createOrUpdateUser($user: UserInput!) {
+        createOrUpdateUser(user: $user) {
+          username
+          id
+        }
+      }
+    `,
+      variables: { user }
+    })
+  })
+
+  const { data } = (await createUserRes.json()) as {
+    data: { createOrUpdateUser: { username: string; id: string } }
+  }
+  const userToken = await getToken(data.createOrUpdateUser.username, 'test')
+
+  const res = await fetch('http://localhost:7070/graphql', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${userToken}`,
+      'x-correlation': `createuser-${firstName}-${familyName}`
+    },
+    body: JSON.stringify({
+      query: `
+      mutation activateUser($userId: String!, $password: String!, $securityQNAs: [SecurityQuestionAnswer]!) {
+        activateUser(userId: $userId, password: $password, securityQNAs: $securityQNAs)
+      }
+    `,
+      variables: {
+        userId: data.createOrUpdateUser.id,
+        password: 'test',
+        securityQNAs: []
+      }
+    })
+  })
+
+  if (res.status !== 200) {
+    console.log(res)
+
+    throw new Error('Could not activate user')
+  }
+
+  return {
+    ...data.createOrUpdateUser,
+    token: userToken,
+    primaryOfficeId,
+    stillInUse: true,
+    password: 'test',
+    isSystemUser: false
+  }
+}
+
+export async function createSystemClient(
+  token: string,
+  officeId: string,
+  scope: 'HEALTH' | 'NATIONAL_ID' | 'EXTERNAL_VALIDATION' | 'AGE_CHECK'
+): Promise<User> {
+  const systemAdmin = await createUser(token, officeId, {
+    role: 'LOCAL_SYSTEM_ADMIN'
+  })
+
+  const createUserRes = await fetch(
+    'http://localhost:3030/registerSystemClient',
+    {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${systemAdmin.token}`,
+        'x-correlation': `create-system-scope`
+      },
+      body: JSON.stringify({ scope })
+    }
+  )
+
+  const credentials: {
+    client_id: string
+    client_secret: string
+    sha_secret: string
+  } = await createUserRes.json()
+
+  const systemToken = await getTokenForSystemClient(
+    credentials.client_id,
+    credentials.client_secret
+  )
+
+  return {
+    token: systemToken,
+    username: credentials.client_id,
+    password: credentials.client_secret,
+    stillInUse: true,
+    primaryOfficeId: officeId,
+    isSystemUser: true
+  }
+}
+
+export async function createUsers(
+  token: string,
+  location: Location,
+  config: Config
+) {
+  const fieldAgents = []
+  const hospitals = []
+  const registrationAgents = []
+  const registrars = []
+  const crvsOffices = (await getFacilities(token))
+    .filter(({ type }) => type === 'CRVS_OFFICE')
+    .filter(({ partOf }) => partOf === 'Location/' + location.id)
+  if (crvsOffices.length === 0) {
+    throw new Error(`Cannot find any CRVS offices for ${location.name}`)
+  }
+  const randomOffice =
+    crvsOffices[Math.floor(Math.random() * crvsOffices.length)]
+  log('Creating field agents')
+  for (let i = 0; i < config.fieldAgents; i++) {
+    fieldAgents.push(
+      await createUser(token, randomOffice.id, {
+        role: 'FIELD_AGENT'
+      })
+    )
+  }
+  log('Field agents created')
+  log('Creating hospitals')
+  for (let i = 0; i < config.hospitalFieldAgents; i++) {
+    hospitals.push(await createSystemClient(token, randomOffice.id, 'HEALTH'))
+  }
+
+  log('Hospitals created')
+  log('Creating registration agents')
+  for (let i = 0; i < config.registrationAgents; i++) {
+    registrationAgents.push(
+      await createUser(token, randomOffice.id, {
+        role: 'REGISTRATION_AGENT'
+      })
+    )
+  }
+  log('Registration agents created')
+  log('Creating local registrars')
+
+  for (let i = 0; i < config.localRegistrars; i++) {
+    registrars.push(
+      await createUser(token, randomOffice.id, {
+        role: 'LOCAL_REGISTRAR'
+      })
+    )
+  }
+  log('Local registrars created')
+
+  return { fieldAgents, hospitals, registrationAgents, registrars }
+}
